Rename index files for DefitionList and FieldList
