+++ src/components/ComponentFactory.js
@@ -15,7 +15,7 @@ import Container from './Container';
 import CSSClass from './CSSClass';
 import CTA from './CTA';
 import CTABanner from './Banner/CTABanner';
-import DefinitionList from './DefinitionList/DefinitionList';
+import DefinitionList from './DefinitionList';
 import DefinitionListItem from './DefinitionList/DefinitionListItem';
 import DeprecatedVersionSelector from './DeprecatedVersionSelector';
 import Describe from './Describe';
@@ -23,7 +23,7 @@ import DriversIndexTiles from './DriversIndexTiles';
 import Emphasis from './Emphasis';
 import Extract from './Extract';
 import Field from './FieldList/Field';
-import FieldList from './FieldList/FieldList';
+import FieldList from './FieldList';
 import Figure from './Figure';
 import Footnote from './Footnote';
 import FootnoteReference from './Footnote/FootnoteReference';
