+++ tests/unit/DefinitionList.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import DefinitionList from '../../src/components/DefinitionList/DefinitionList';
+import DefinitionList from '../../src/components/DefinitionList';
 import mockData from './data/DefinitionList.test.json';
 
 it('DefinitionList renders correctly', () => {
