+++ tests/unit/FieldList.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import FieldList from '../../src/components/FieldList/FieldList';
+import FieldList from '../../src/components/FieldList';
 import { theme } from '../../src/theme/docsTheme';
 import { ThemeProvider } from '@emotion/react';
 // data for this component
